Simplify and clean code, needs further work
