+++ lib/finance.dart
@@ -1,21 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-
-const String currency = '€';
-
-class Move {
-  final String descriptor;
-  final int balance;
-  const Move({required this.descriptor, required this.balance});
-}
-
-int calculateTotalFinancialMoves(List<Move> movesList) {
-  int balance = 0;
-  for (var move in movesList) {
-    balance += move.balance;
-  }
-  return balance;
-}
+import 'package:money_manager/types/finance.dart';
 
 class FinancialModel extends ChangeNotifier {
   final BuildContext context;
@@ -27,7 +12,8 @@ class FinancialModel extends ChangeNotifier {
   List<Move> get items => _items;
 
   /// The current total price of all items
-  int get total => calculateTotalFinancialMoves(items);
+  int get total =>
+      calculateTotalMoves(items.map<int>((e) => e.balance).toList());
 
   /// Adds [item] to cart
   void add(Move item) {
@@ -47,7 +33,8 @@ class DebtModel extends ChangeNotifier {
   final BuildContext context;
   final List<Move> _items = [];
   List<Move> get items => _items;
-  int get total => calculateTotalFinancialMoves(items);
+  int get total =>
+      calculateTotalMoves(items.map<int>((e) => e.balance).toList());
 
   void add(Move item) {
     _items.add(item);
@@ -76,7 +63,8 @@ class BorrowModel extends ChangeNotifier {
   final BuildContext context;
   final List<Move> _items = [];
   List<Move> get items => _items;
-  int get total => calculateTotalFinancialMoves(items);
+  int get total =>
+      calculateTotalMoves(items.map<int>((e) => e.balance).toList());
 
   void add(Move item) {
     // Add negative balance to balance, add to borrow; then clear out
