+++ lib/savedata/native.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'dart:convert';
 import 'package:path/path.dart';
 import 'package:path_provider/path_provider.dart';
-import 'package:money_manager/savedata.dart';
+import 'package:money_manager/types/savedata.dart';
 
 class PlatformSaveData extends SaveData {
   Future<String> getPathToSaveData() async {
