+++ lib/widgets/main_navigation_bar.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+
+class MainNavigationBar extends StatelessWidget {
+  const MainNavigationBar(
+      {super.key,
+      required this.selectedIndex,
+      required this.onDestinationSelected});
+
+  final int selectedIndex;
+  final Function(int) onDestinationSelected;
+
+  @override
+  Widget build(BuildContext context) {
+    return NavigationBar(
+      destinations: const [
+        NavigationDestination(
+            icon: Icon(Icons.account_balance), label: "Balance"),
+        NavigationDestination(icon: Icon(Icons.person), label: "Debt"),
+        NavigationDestination(icon: Icon(Icons.attach_money), label: "Borrow")
+      ],
+      selectedIndex: selectedIndex,
+      onDestinationSelected: onDestinationSelected,
+    );
+  }
+}
